Add flag to  override the occi location hostname.
If the OCCI WSGI server is running behind a proxy, the location is
incorrectly set. This flags allows to override the hostname of the
location with a custom value.

Note that although the name is "hostname" (from pyssf) the correct value
for that field is something like http://myocciserver.example.org:PORT